+++ lib/rails_api_format/lib/rails_api_format/devise_failure_app.rb
@@ -1,7 +1,0 @@
-module RailsApiFormat
-  class DeviseFailureApp < Devise::FailureApp
-    def http_auth_body
-      Error.new(status: :unauthorized, error: i18n_message).to_json
-    end
-  end
-end